libc: Update
CMake now uses toolchain file instead of requiring toolchain in PATH.

`exit()` has been implemented, now it's time to implement it as a syscall.
