+++ user/libc/unistd.h
@@ -1,7 +1,7 @@
 #ifndef _UNISTD_H
 #define _UNISTD_H
 
-#include <sys/types.h>
+#include "sys/types.h"
 
 #if defined (__cplusplus)
 extern "C" {
@@ -14,6 +14,8 @@ extern "C" {
 
     ssize_t read(int fd, const void* buffer, size_t count);
     ssize_t write(int fd, const void* buffer, size_t count);
+
+  __attribute__((noreturn)) void exit(int status);
 
 #if defined (__cplusplus)
 } /* extern "C" */
